New script
New script. It assembles flat diffraction patterns in Fourierspace according to where they are on the Ewald spheres.
Comments and instructions are in mixed Swedish and English. A full English version can be created if it is requested by some one.

You are free to use the code and change it, privately. But if you want to publish it or share your changed code or the original code then you have to email me and ask me for permission.